requirements.txt
# Core
#fastapi==0.104.1
#uvicorn==0.24.0
#pydantic
#numpy<2.0
#Pillow==10.4.0
#safetensors==0.4.5

# ML frameworks
#torch==2.0.1+cpu
#torchvision==0.15.2+cpu
#torchaudio==2.0.2+cpu
#--extra-index-url https://download.pytorch.org/whl/cpu

# ModelScope + Diffusers + Transformers
# modelscope==1.11.0
#diffusers==0.27.2
#transformers==4.44.0
#huggingface_hub==0.24.7
#accelerate==0.33.0

# Audio
#scipy==1.10.1
#librosa==0.10.1
#soundfile==0.12.1

# Utilities
#opencv-python==4.8.1.78
#ffmpeg-python

fastapi
uvicorn
torch
torchvision
torchaudio
Pillow

